Learning Project: Start the Week on Thursday - Introducing the Modulo Operator %
Problem Outline: Currently, your script calculates 7 days starting from
today. We want to modify it so that it first figures out the date of
the most recent Thursday (it could be today if today is Thursday, or
it could be last Thursday) and then generates the 7 days starting from
that Thursday.

Left  test bed in to show working.